% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseNest}
\alias{parseNest}
\title{Collapse a nested character list into a character string 'varname=123;othervar=foo', 'varname=0;othervar=bar', ...}
\usage{
parseNest(x, symbol1 = "=", symbol2 = ",", symbol3 = ";")
}
\description{
Description
}
