% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mat_image.R
\name{test.mat.image}
\alias{test.mat.image}
\title{test.mat.image}
\usage{
test.mat.image(
  x,
  candidates.df,
  xlim = c(1, di[2]),
  ylim = c(di[1], 1),
  aspect = "iso",
  sub = sprintf("Dimensions: \%d x \%d", di[1], di[2]),
  xlab = "Column",
  ylab = "Row",
  cuts = 15,
  useRaster = FALSE,
  useAbs = NULL,
  colorkey = !useAbs,
  col.regions = NULL,
  lwd = NULL,
  border.col = NULL,
  ...
)
}
\arguments{
\item{x}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
c()


}
