% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-best-flags.R
\name{tf_build_buildings}
\alias{tf_build_buildings}
\title{tf_build_buildings}
\usage{
tf_build_buildings(
  url.townforged,
  url.wallet,
  flags.to.buy.df,
  build.tx.hashes.v,
  building.type,
  economic.power,
  coords.origin,
  city
)
}
\arguments{
\item{url.townforged}{TODO}

\item{url.wallet}{TODO}

\item{flags.to.buy.df}{TODO}

\item{building.type}{TODO}

\item{economic.power}{TODO}

\item{coords.origin}{TODO}

\item{city}{TODO}
}
\value{
TODO
}
\description{
Buys flags
}
\examples{
c()


}
