% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-best-flags.R
\name{tf_buy_flags}
\alias{tf_buy_flags}
\title{tf_buy_flags}
\usage{
tf_buy_flags(url.wallet, flags.to.buy.df, coords.origin, city)
}
\arguments{
\item{url.wallet}{TODO}

\item{flags.to.buy.df}{TODO}

\item{coords.origin}{TODO}

\item{city}{TODO}
}
\value{
TODO
}
\description{
Buys flags
}
\examples{
c()


}
