% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-best-flags.R
\name{tf_expand_search_flag}
\alias{tf_expand_search_flag}
\title{tf_expand_search_flag}
\usage{
tf_expand_search_flag(
  url.rpc,
  flag.bounds.ls,
  infl.grid.disagg,
  west.coord,
  south.coord,
  east.coord,
  north.coord,
  init.lat,
  init.long,
  role.id,
  economic.power,
  city,
  subject.to.influence,
  get.flag.cost,
  infl.thresholds
)
}
\arguments{
\item{url.rpc}{TODO}
}
\value{
TODO
}
\description{
Expands a flag as tf_search_best_flags() searches
}
\examples{
c()


}
