% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-best-flags.R
\name{tf_get_best_flag_map}
\alias{tf_get_best_flag_map}
\title{tf_get_best_flag_map}
\usage{
tf_get_best_flag_map(
  url,
  candidates.df,
  chosen.item.id,
  number.of.top.candidates,
  display.perimeter = TRUE
)
}
\arguments{
\item{url}{TODO}

\item{candidates.df}{TODO}

\item{chosen.item.id}{TODO}

\item{number.of.top.candidates}{TODO}

\item{display.perimeter}{TODO}
}
\value{
TODO
}
\description{
Calculates matrix of best flag(s) map
}
\examples{
c()


}
