% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-game-mechanics.R
\name{tf_get_influence_effects}
\alias{tf_get_influence_effects}
\title{Construct influence effects matrix data from Townforge C++ file}
\usage{
tf_get_influence_effects(
  source.dir = "https://git.townforge.net/townforge/townforge/raw/branch/cc",
  ...
)
}
\arguments{
\item{source.dir}{directory of Townforge source}

\item{...}{TODO}
}
\description{
Description
}
\details{
Construct influence effects matrix
}
