% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-game-mechanics.R
\name{tf_get_min_flag_size_data}
\alias{tf_get_min_flag_size_data}
\title{Construct minimum flag size data from Townforge C++ file}
\usage{
tf_get_min_flag_size_data(
  source.dir = "https://git.townforge.net/townforge/townforge/raw/branch/cc",
  ...
)
}
\arguments{
\item{source.dir}{directory of Townforge source}

\item{...}{TODO}
}
\description{
Description
}
\details{
Construct minimum flag size data
}
