% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-game-mechanics.R
\name{tf_get_seed_words_list}
\alias{tf_get_seed_words_list}
\title{Gather English wallet mnemonic seed words list}
\usage{
tf_get_seed_words_list(
  source.url =
    "https://raw.githubusercontent.com/monero-project/monero/master/src/mnemonics/english.h",
  ...
)
}
\arguments{
\item{source.url}{directory of Townforge source}

\item{...}{TODO}
}
\description{
Description
}
\details{
Construct influence effects matrix
}
