% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{tf_infl_grid}
\alias{tf_infl_grid}
\title{tf_infl_grid}
\usage{
tf_infl_grid(
  url,
  building.type,
  effect.type,
  grid.dim,
  coords.origin,
  disaggregated = FALSE
)
}
\arguments{
\item{url}{TODO}

\item{building.type}{TODO}

\item{effect.type}{TODO}

\item{disaggregated}{return object disaggregates matrix by building type}
}
\value{
TODO
}
\description{
Get boundaries of building influence
}
\examples{
c()


}
