% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{tf_min_flag_size}
\alias{tf_min_flag_size}
\title{tf_min_flag_size}
\usage{
tf_min_flag_size(building.role, economic.power)
}
\arguments{
\item{building.role}{TODO}

\item{economic.power}{TODO}

\item{min.size.scale.df}{TODO}
}
\value{
TODO
}
\description{
Get boundaries of building influence
}
\examples{
c()


}
