% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tf_parse_accounts}
\alias{tf_parse_accounts}
\title{tf_parse_accounts}
\usage{
tf_parse_accounts(
  ids,
  get_accounts = "cc_get_accounts",
  get_account = "cc_get_account",
  fields_plain = c("name", "public_key", "moose_killed", "bears_killed", "prestige",
    "description", "balance"),
  fields_nests = c("flags", "item_balances", "badges"),
  collapse_fun = TownforgeR:::parseNest,
  truncate_char = 40,
  url.rpc = "http://127.0.0.1:18881/json_rpc",
  ...
)
}
\arguments{
\item{ids}{TODO}

\item{get_accounts}{TODO}

\item{get_account}{TODO}

\item{fields_plain}{TODO}

\item{fields_nests}{TODO}

\item{collapse_fun}{TODO}

\item{truncate_char}{TODO}

\item{url.rpc}{TODO}

\item{...}{TODO}
}
\value{
A data.frame corresponding to queried plain and nested fields over desired ids
}
\description{
Parse existing user accounts and some key info into an R data.frame
}
\examples{
tf_parse_accounts()

}
