% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tf_parse_markets}
\alias{tf_parse_markets}
\title{tf_parse_markets}
\usage{
tf_parse_markets(
  get_order_book = "cc_get_order_book",
  url.rpc = "http://127.0.0.1:18881/json_rpc"
)
}
\arguments{
\item{get_order_book}{TODO}

\item{url.rpc}{TODO}
}
\description{
Description
}
