% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tf_parse_network}
\alias{tf_parse_network}
\title{tf_parse_network}
\usage{
tf_parse_network(
  get_info = "get_info",
  url = "http://127.0.0.1:18881/json_rpc"
)
}
\arguments{
\item{get_info}{TODO}

\item{url}{TODO}
}
\description{
Creates verbatim output string that describes Townforge network status
}
