% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tf_parse_nfts}
\alias{tf_parse_nfts}
\title{tf_parse_nfts}
\usage{
tf_parse_nfts(get_custom_items = "cc_get_custom_items", truncate_char = 80)
}
\arguments{
\item{get_custom_items}{TODO}

\item{truncate_char}{TODO}
}
\description{
Description
}
