% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{tf_plot_influence}
\alias{tf_plot_influence}
\title{tf_plot_influence}
\usage{
tf_plot_influence(url, building.type, effect.type, cut.out.flags = TRUE)
}
\arguments{
\item{url}{TODO}

\item{building.type}{TODO}

\item{effect.type}{TODO}

\item{cut.out.flags}{TODO}
}
\value{
TODO
}
\description{
Plot buildings' influence map
}
\examples{
c()


}
