% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{tf_rpc_curl}
\alias{tf_rpc_curl}
\title{TownforgeR function for curl'ing RPC calls}
\usage{
tf_rpc_curl(
  url = "http://127.0.0.1:18881/json_rpc",
  method = "get_block_count",
  params = list(),
  ...
)
}
\description{
Description
}
