% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-best-flags.R
\name{tf_search_best_flags}
\alias{tf_search_best_flags}
\title{tf_search_best_flags}
\usage{
tf_search_best_flags(
  url.rpc,
  building.type,
  economic.power,
  get.flag.cost = TRUE,
  city,
  grid.density.params = c(10, 10),
  in.shiny = FALSE,
  waitress = NULL
)
}
\arguments{
\item{url.rpc}{TODO}

\item{building.type}{TODO}

\item{economic.power}{TODO}

\item{get.flag.cost}{TODO}

\item{city}{TODO}

\item{grid.density.params}{TODO}

\item{in.shiny}{TODO}

\item{waitress}{TODO}
}
\value{
TODO
}
\description{
Brute force search for flags with best production and return on investment
f
}
\examples{
c()


}
