% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-wallet.R
\name{tf_server_wallet_load}
\alias{tf_server_wallet_load}
\title{tf_server_wallet_load}
\usage{
tf_server_wallet_load(wallet.username, wallet.password, wallet.directory)
}
\arguments{
\item{wallet.username}{TODO}

\item{wallet.password}{TODO}

\item{wallet.directory}{TODO}
}
\value{
TODO
}
\description{
Load a wallet with new townforge-wallet-rpc instance
}
\examples{
c()


}
