% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-wallet.R
\name{tf_server_wallet_restore}
\alias{tf_server_wallet_restore}
\title{tf_server_wallet_restore}
\usage{
tf_server_wallet_restore(
  wallet.restore.username,
  wallet.restore.password,
  wallet.restore.seed.words,
  wallet.restore.seed.offset.passphrase,
  wallet.directory
)
}
\arguments{
\item{wallet.restore.username}{TODO}

\item{wallet.restore.password}{TODO}

\item{wallet.restore.seed.words}{TODO}

\item{wallet.restore.seed.offset.passphrase}{TODO}

\item{wallet.directory}{TODO}
}
\value{
TODO
}
\description{
Create a wallet file from seed words via townforge-wallet-cli. Expected to be used on a server
}
\examples{
c()


}
