% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tf_shiny_nft_png}
\alias{tf_shiny_nft_png}
\title{tf_shiny_nft_png}
\usage{
tf_shiny_nft_png(item_hash = "", storage_blocks = 2.5 * 24 * 60)
}
\arguments{
\item{item_hash}{TODO}
}
\description{
Description
}
