% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\docType{data}
\name{uiTF}
\alias{uiTF}
\title{Shiny UI for browser}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
uiTF
}
\description{
Description
}
\keyword{datasets}
