% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{tf_flag_bounds}
\alias{tf_flag_bounds}
\title{tf_flag_bounds}
\usage{
tf_flag_bounds(url, grid.dim, coords.offset = 1000)
}
\arguments{
\item{url}{TODO}

\item{grid.dim}{TODO}

\item{coords.offset}{TODO}
}
\value{
TODO
}
\description{
Get boundaries of flags
}
\examples{
c()


}
