% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{tf_infl_location}
\alias{tf_infl_location}
\title{tf_infl_location}
\usage{
tf_infl_location(
  url.rpc,
  building.type = "all",
  coords.origin = NULL,
  grid.dim = NULL,
  coords.offset = 1000
)
}
\arguments{
\item{url.rpc}{TODO}

\item{building.type}{TODO}

\item{coords.origin}{TODO}

\item{grid.dim}{TODO}

\item{coords.offset}{TODO}
}
\value{
TODO
}
\description{
Get boundaries of building influence
}
\examples{
c()


}
