% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{tf_parse_nfts}
\alias{tf_parse_nfts}
\title{tf_parse_nfts}
\usage{
tf_parse_nfts(
  get_custom_items = "cc_get_custom_items",
  truncate_char = 80,
  url = "http://127.0.0.1:18881/json_rpc"
)
}
\arguments{
\item{get_custom_items}{TODO}

\item{truncate_char}{TODO}

\item{url}{TODO}
}
\description{
Description
}
