% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpc.R
\name{tf_rpc_curl}
\alias{tf_rpc_curl}
\title{TownforgeR function for curl'ing RPC calls}
\usage{
tf_rpc_curl(
  url = "http://127.0.0.1:18881/json_rpc",
  method = "get_block_count",
  params = list(),
  num.as.string = FALSE,
  nonce.as.string = FALSE,
  ...
)
}
\arguments{
\item{url}{TODO}

\item{method}{One of: '}

\item{params}{TODO}

\item{num.as.string}{TODO}

\item{nonce.as.string}{TODO}

\item{...}{TODO}
}
\description{
Description
}
\details{
A list of RPC commands compatible with Townforge (v0.27.0.7 at the time of writing this) is available via Python code in https://git.townforge.net/townforge/townforge/src/branch/cc/utils/python-rpc/framework/daemon.py
}
