% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-best-flags.R
\name{tf_search_best_flags}
\alias{tf_search_best_flags}
\title{tf_search_best_flags}
\usage{
tf_search_best_flags(
  url,
  building.type,
  economic.power,
  get.flag.cost = TRUE,
  city = 0,
  grid.density.params = c(10, 10)
)
}
\arguments{
\item{url}{TODO}

\item{economic.power}{TODO}

\item{get.flag.cost}{TODO}

\item{city}{TODO}

\item{grid.density.params}{TODO}
}
\value{
TODO
}
\description{
Brute force search for flags with best production and return on investment
}
\examples{
c()


}
